deploy_manager/requirements.txt
boto3>=1.34.0
click>=8.1.0
rich>=13.7.0
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
moto>=4.2.0
